be/chat-service/chatbot/ai.py
import numpy as np
import tensorflow as tf

# Training data
X_train = np.array([
    [1, 1, 0, 1, 0, 0],  # Flu
    [0, 1, 1, 0, 0, 0],  # Cold
    [1, 1, 0, 0, 1, 0],  # COVID-19
    [0, 0, 1, 0, 0, 1]   # Allergy
], dtype=np.float32)
y_train = tf.keras.utils.to_categorical([0, 1, 2, 3], num_classes=4)
diseases = ["Flu", "Cold", "COVID-19", "Allergy"]

def build_model():
    inputs = tf.keras.Input(shape=(6,))
    x = tf.keras.layers.Dense(16, activation='relu')(inputs)
    x = tf.keras.layers.Dropout(0.5)(x, training=True)
    x = tf.keras.layers.Dense(16, activation='relu')(x)
    x = tf.keras.layers.Dropout(0.5)(x, training=True)
    outputs = tf.keras.layers.Dense(4, activation='softmax')(x)
    return tf.keras.Model(inputs, outputs)

model = build_model()
model.compile(optimizer='adam', loss='categorical_crossentropy', metrics=['accuracy'])
model.fit(X_train, y_train, epochs=100, verbose=0)

def predict_with_uncertainty(input_array, n_iter=100):
    preds = np.array([model(input_array, training=True).numpy() for _ in range(n_iter)])
    mean = preds.mean(axis=0)
    std = preds.std(axis=0)
    return mean, std

test_map = {
    "Flu": "Influenza A/B test",
    "Cold": "Nasal swab",
    "COVID-19": "PCR test",
    "Allergy": "Allergy skin test"
}
medicine_map = {
    "Flu": "Oseltamivir (Tamiflu)",
    "Cold": "Rest, fluids, antihistamines",
    "COVID-19": "Isolation + Paracetamol",
    "Allergy": "Loratadine or Cetirizine"
}
