be/chat-service/requirements.txt
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
click==8.2.0
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.2
cryptography==45.0.2
cycler==0.12.1
daphne==4.2.0
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
fonttools==4.58.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
joblib==1.5.0
kiwisolver==1.4.8
matplotlib==3.10.3
msgpack==1.1.0
nltk==3.9.1
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pillow==11.2.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
redis==6.1.0
regex==2024.11.6
scikit-learn==1.6.1
scipy==1.15.3
service-identity==24.2.0
setuptools==80.7.1
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tqdm==4.67.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
tzdata==2025.2
zope.interface==7.2
